requirements/base.txt
Django==2.0.4
pytz==2018.4
psycopg2==2.7.3.1
celery==4.1.0
redis==2.10.5
django-filter==1.1.0
django-rest-swagger==2.1.2
djangorestframework==3.8.2
django-solo==1.1.3